Improve auth code structure
